backend/requirements.txt
dj-database-url==0.5.0
channels==3.0.4
pyqrcode
qrcode
channels-redis
pycrypto
django==4.0.0
django-admin-thumbnails==0.2.6
django-baton==2.3.0
django-cas-ng==4.2.1
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-dbconn-retry==0.1.5
django-dramatiq==0.10.0
django-import-export==2.8.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-redis==5.0.0
django-rest-swagger==2.1.2
django-simple-captcha==0.5.17
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
otpauth==1.0.1
Pillow==8.4.0
pip==22.2.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.26.0
django-jazzmin
dj_database_url
pytz
celery